Separate users reducer into 2 different reducers for performance
