+++ src/screens/HomeScreen/MatchesList.js
@@ -1,18 +1,14 @@
-import React, { Component, useState, useEffect } from 'react';
-import { useSelector, useDispatch, connect } from 'react-redux';
-import { SafeAreaView, FlatList, View, Text } from 'react-native';
-import { fetchPotentialMatches } from '../../store/usersReducer';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { SafeAreaView, FlatList } from 'react-native';
+import { fetchPotentialMatches } from '../../store/potentialMatchesReducer';
 import styles from './styles';
 import UserRow from './UserRow';
-import SearchBar from '../../components/SearchBar';
 import { firebase } from '../../firebaseSpecs/config';
 
 export class MatchesList extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-      matchedUsers: [],
-    };
   }
 
   componentDidMount() {
@@ -22,7 +18,7 @@ export class MatchesList extends Component {
   }
 
   render() {
-    const { user, users, navigation } = this.props;
+    const { user, potentialMatches, navigation } = this.props;
     const currentUserLikes = user.likes;
 
     const renderItem = ({ item }) => (
@@ -30,7 +26,9 @@ export class MatchesList extends Component {
     );
 
     // look through current user's likes array & find matches
-    const matches = users.filter((user) => currentUserLikes.includes(user.id));
+    const matches = potentialMatches.filter((user) =>
+      currentUserLikes.includes(user.id)
+    );
 
     if (matches.length > 0) {
       return (
@@ -51,7 +49,7 @@ const mapState = (state) => ({
   // logged-in user
   user: state.user,
   // users that are potential matches
-  users: state.users,
+  potentialMatches: state.potentialMatches,
 });
 
 const mapDispatch = (dispatch) => ({
