+++ src/store/discoverUsersReducer.js
@@ -1,15 +1,9 @@
 import { firebase } from '../firebaseSpecs/config';
 
 const SET_DISCOVER_USERS = 'SET_ALL_USERS';
-const SET_POTENTIAL_MATCHES = 'SET_POTENTIAL_MATCHES';
 
 export const setDiscoverUsers = (users) => ({
   type: SET_DISCOVER_USERS,
-  users,
-});
-
-export const setPotentialMatches = (users) => ({
-  type: SET_POTENTIAL_MATCHES,
   users,
 });
 
@@ -30,32 +24,9 @@ export const fetchUsersWithInterests = (interests) => {
   };
 };
 
-export const fetchPotentialMatches = (id) => {
-  return async (dispatch) => {
-    const userRef = firebase.firestore().collection('users');
-    await userRef
-      .where('likes', 'array-contains', id)
-      .get()
-      .then((snapshot) => {
-        if (snapshot) {
-          let users = snapshot.docs.map((doc) => {
-            const data = doc.data();
-            const id = doc.id;
-            return { id, ...data };
-          });
-          dispatch(setPotentialMatches(users));
-        } else {
-          console.log('users do not exist');
-        }
-      });
-  };
-};
-
 export default function (state = [], action) {
   switch (action.type) {
     case SET_DISCOVER_USERS:
-      return action.users;
-    case SET_POTENTIAL_MATCHES:
       return action.users;
     default:
       return state;
